+++ src/led/shader.rs
@@ -3,28 +3,28 @@ use vecmath::*;
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 pub struct LightEnvironment {
 	pub winches: Vec<WinchLighting>,
-	pub flash_exponent: f64,
-	pub flash_rate_hz: f64,
-	pub winch_wave_exponent: f64,
-	pub winch_wavelength: f64,
-	pub winch_wave_window_length: f64,
-	pub winch_command_color: Vector3<f64>,
-	pub winch_motion_color: Vector3<f64>,
-	pub brightness: f64,
+	pub flash_exponent: f32,
+	pub flash_rate_hz: f32,
+	pub winch_wave_exponent: f32,
+	pub winch_wavelength: f32,
+	pub winch_wave_window_length: f32,
+	pub winch_command_color: Vector3<f32>,
+	pub winch_motion_color: Vector3<f32>,
+	pub brightness: f32,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 pub struct WinchLighting {
-	pub command_phase: f64,
-	pub motion_phase: f64,
-	pub wave_amplitude: f64,
-	pub base_color: Vector3<f64>,
-	pub flash_color: Vector3<f64>,
+	pub command_phase: f32,
+	pub motion_phase: f32,
+	pub wave_amplitude: f32,
+	pub base_color: Vector3<f32>,
+	pub flash_color: Vector3<f32>,
 }
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct PixelMapping {
-	pub location: Vector3<f64>,
+	pub location: Vector3<f32>,
 	pub usage: PixelUsage,
 }
 
@@ -37,11 +37,11 @@ pub enum PixelUsage {
 
 #[derive(Debug)]
 pub struct Shader {
-	pub flash_state_radians: f64,
-	pub temp_spin_radians: f64,
+	pub flash_state_radians: f32,
+	pub temp_spin_radians: f32,
 }
 
-fn pulse_wave(angle: f64, exponent: f64) -> f64 {
+fn pulse_wave(angle: f32, exponent: f32) -> f32 {
 	(0.5 + 0.5 * angle.sin()).powf(exponent)
 }
 
@@ -53,17 +53,17 @@ impl Shader {
 		}
 	}
 
-	pub fn step(&mut self, env: &LightEnvironment, seconds: f64) {
+	pub fn step(&mut self, env: &LightEnvironment, seconds: f32) {
 		self.flash_state_radians = (self.flash_state_radians + seconds * env.flash_rate_hz * TAU) % TAU;
-		const SPIN_HZ : f64 = 0.5;
+		const SPIN_HZ : f32 = 0.5;
 		self.temp_spin_radians = (self.temp_spin_radians + seconds * SPIN_HZ * TAU) % TAU;
 	}
 
-	fn flash(&self, env: &LightEnvironment) -> f64 {
+	fn flash(&self, env: &LightEnvironment) -> f32 {
 		pulse_wave(self.flash_state_radians, env.flash_exponent)
 	}
 
-	fn winch_wave(&self, winch: &WinchLighting, env: &LightEnvironment, map: &PixelMapping, phase: f64) -> f64 {
+	fn winch_wave(&self, winch: &WinchLighting, env: &LightEnvironment, map: &PixelMapping, phase: f32) -> f32 {
 		let z = map.location[2];
 		let pulse_theta = phase + z * TAU / env.winch_wavelength;
 		let window_theta = z * TAU / env.winch_wave_window_length;
@@ -71,7 +71,7 @@ impl Shader {
 		winch.wave_amplitude * window * pulse_wave(pulse_theta, env.winch_wave_exponent)
 	}
 
-	pub fn pixel(&self, env: &LightEnvironment, map: &PixelMapping) -> Vector3<f64> {
+	pub fn pixel(&self, env: &LightEnvironment, map: &PixelMapping) -> Vector3<f32> {
 		let c = match map.usage {
 
 			PixelUsage::Winch(id) => {
